Created all the needed subscription requests and model classes - Refactoring is needed

+++ quickpaysdk/src/main/java/net/quickpay/quickpaysdk/networking/quickpayapi/quickpaylink/models/QPBasket.kt
@@ -5,7 +5,7 @@ import org.json.JSONObject
 class QPBasket(qty: Int,
                item_no: String,
                item_name: String,
-               item_price: Int,
+               item_price: Double,
                vat_rate: Double): JSONObject() {
 
     // Required properties
@@ -13,7 +13,7 @@ class QPBasket(qty: Int,
     var qty: Int = qty
     var item_no: String = item_no
     var item_name: String = item_name
-    var item_price: Int = item_price
+    var item_price: Double = item_price
     var vat_rate: Double = vat_rate
 
 }
